requirements.txt
openai-whisper
openai
python-dotenv
pydub
tqdm